app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.26.3
packaging==23.1
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.3.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
psycopg2==2.9.6
pure-eval==0.2.2
pydantic==2.0.3
pydantic_core==2.3.0
pydantic-settings==2.0.3
Pygments==2.16.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
wheel==0.38.4
yarl==1.9.2
zipp==3.16.2